pipeline/requirements.txt
autopep8==2.0.2
category_encoders==2.6.1
dacite==1.8.0
joblib==1.2.0
mord==0.6
numpy==1.24.3
pandas==2.0.1
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
tzdata==2023.3
imbalanced-learn==0.10.1
